pyproject.toml
[tool.poetry]
authors = ["Manuel Lera Ramirez <manulera14@gmail.com>"]
description = ""
name = "allele_qc"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.9"
biopython = "^1.79"
pydantic = "^1.10.2"
pandas = "^1.5.0"
openpyxl = "^3.0.10"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
autopep8 = "^1.7.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
